requirements.txt
requests>=2.32.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
openai>=1.0.0,<2.0.0
